Corrected retry handling
Previous:
If the requests returns an unwanted response code we attempt to retry.
We will (most likely) always receive the exact same response code every
single attempt, which will lead to the operation failing with error timeout
or 'unexpectedResponseCode' (which we would have received much earlier).

Current:
A retry should only be performed if the request fails to be sent or if a response is never received.